+++ scripts/plot_cov_alignment.py
@@ -14,7 +14,7 @@ from param_utils import Params
 
 if __name__ == '__main__':
     parser = argparse.ArgumentParser()
-    parser.add_argument('--covrot', type=float, default=60.0,
+    parser.add_argument('--covrot', type=int, default=60,
                         help='Rotation angle to create a distribution')
     args = parser.parse_args()
